vpt: Minor renamings and cleanups.
